requirements.txt
asgiref==3.3.1
astroid==2.4.2
colorama==0.4.4
Django==3.1.5
django-cors-headers==3.6.0
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
Faker==5.0.1
gunicorn==20.0.4
isort==5.7.0
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
Pillow==8.1.0
psycopg2-binary==2.8.6
pylint==2.6.0
pylint-django==2.4.1
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
wrapt==1.12.1
